refactor: move `get_interface_file_path` to avoid circular import
